requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.8
django-betterforms==1.2
django-crispy-forms==1.9.1
django-filter==2.3.0
django-tabular-permissions==2.6
djangorestframework==3.11.0
djangorestframework-datatables==0.5.2
googlemaps==4.4.1
idna==2.10
Pillow==7.2.0
psycopg2==2.8.5
python-decouple==3.3
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.9
